pyproject.toml
[project]
name = "mirrornote-diarization-mlx"
version = "0.1.0"
description = "Local speaker diarization parity experiments for MirrorNote"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
]

[project.optional-dependencies]
audio = [
  "soundfile>=0.12",
]
pyannote = [
  "numpy<2",
  "pyannote.audio==3.1.1",
  "torch==2.2.2",
  "torchaudio==2.2.2",
]
mlx = [
  "mlx>=0.20",
]
dev = [
  "pytest>=8",
  "jsonschema>=4.22",
]

[project.scripts]
mirrornote-diarize = "mirrornote_diarization.segmentation_parity:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mirrornote_diarization"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
